pyproject.toml
[tool.poetry]
name = "collect_service"
version = "0.1.0"
description = ""
authors = ["oleg"]

[tool.poetry.dependencies]
python = "^3.9"
appdirs = "^1.4.4"
asgiref = "^3.4.0"
attrs = "^21.2.0"
autopep8 = "^1.5.7"
certifi = "^2021.5.30"
charset-normalizer = "^2.0.1"
distlib = "^0.3.2"
Django = "^3.2.4"
django-extensions = "^3.1.3"
django-nine = "^0.2.4"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.1"
filelock = "^3.0.12"
gprof2dot = "^2021.2.21"
idna = "^3.2"
inflection = "^0.5.1"
iniconfig = "^1.1.1"
Jinja2 = "^3.0.1"
MarkupSafe = "^2.0.1"
packaging = "^20.9"
Pillow = "^8.2.0"
pluggy = "^0.13.1"
psycopg2-binary = "^2.9.1"
py = "^1.10.0"
pycodestyle = "^2.7.0"
Pygments = "^2.9.0"
PyJWT = "^2.1.0"
pyparsing = "^2.4.7"
python-dateutil = "^2.8.1"
pytz = "^2021.1"
requests = "^2.26.0"
six = "^1.16.0"
sqlparse = "^0.4.1"
text-unidecode = "^1.3"
toml = "^0.10.2"
urllib3 = "^1.26.6"
virtualenv = "^20.4.7"
black = "^21.7b0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.0"
Faker = "^8.9.1"
freezegun = "^1.1.0"
django-shell-plus = "^1.1.7"
pytest = "^6.2.4"
pytest-django = "^4.4.0"
pytest-factoryboy = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
